+++ NEWS.md
@@ -9,8 +9,11 @@ enrichment and visualize respective results
     - Keep selected timepoint, dosage and cell line options when selecting
     another perturbation type
     - Add bubble plot of CMap perturbation types
-* `launchResultPlotter()` now allows to view tables below specific plots and
-drag-and-select those plots to filter data in those same tables
+* `launchResultPlotter()`:
+    - Now allows to view tables below specific plots and drag-and-select those
+    plots to filter data in those same tables
+    - When plotting targeting drugs and similar perturbations, update available
+    columns and correctly use user-selected column to plot
 * `launchMetadataViewer()` now correctly parses values from `Input`
 attributes as numeric